Separate test definitions from test engine
