+++ include/lhp_test_parse.h
@@ -1,0 +1,21 @@
+/*
+ * File-Name: lhp_test_parser.h
+ * File-Desc: Header file for test parser functions for lh_parser
+ * App-Name: las-header-parser
+ * Project-Name: Las-Header-Parser
+ * Copyright: Copyright (c) 2020, DC Slagel
+ * License-Identifier: MIT
+ */
+
+
+#ifndef LHP_TEST_PARSE_H
+#define LHP_TEST_PARSE_H
+
+int test_parse_mnemonic(void);
+int test_parse_empty_unit(void);
+int test_parse_value(void);
+int test_parse_desc(void);
+int test_parse_m_unit(void);
+int test_parse_empty_desc(void);
+
+#endif /* LHP_TEST_PARSE_H */
